+++ ChangeLog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v1.6.2
+
+- Added Product:Card component (26/07/2025)
+- Added Product:Button component (26/07/2025)
+
 ## v1.6.1
 
 - Added a Slider component to be reused using product.medias (19/07/2025)